+++ src/main/java/roomescape/reservation/application/ReservationService.java
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.EscapeRoomException;
-import roomescape.reservation.Reservation;
-import roomescape.reservation.dto.request.ReservationSaveRequest;
-import roomescape.reservation.dto.response.ReservationDetailFindResponse;
-import roomescape.reservation.dto.response.ReservationSaveResponse;
-import roomescape.reservation.infrastructure.ReservationRepository;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.application.dto.request.ReservationSaveRequest;
+import roomescape.reservation.application.dto.response.ReservationDetailFindResponse;
+import roomescape.reservation.application.dto.response.ReservationSaveResponse;
+import roomescape.reservation.application.port.out.ReservationRepository;
 import roomescape.slot.SlotOccupancy;
 import roomescape.slot.Slot;
 import roomescape.slot.application.SlotAssembler;
