+++ src/test/java/roomescape/LayerSeparationTest.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.config.TestTimeConfig;
-import roomescape.reservation.presentation.UserReservationController;
+import roomescape.reservation.adapter.in.web.UserReservationController;
 
 import java.lang.reflect.Field;
